Add Host_DFU_RN-BM as individual files.
This commit fixes issue #1. Instead of misusing a .7z archive for this
purpose, the source code is added as individual files to the repository.

Artifacts which are the result of the build process are removed from
the repository.

Only PreMPTool_new.exe is kept here as it seems to be required, and
because it is referenced from the README.md. Actually, it should go into
a tool directory.